Add bled112 support for scanning and device discovery.
Added BLED112Backend.scan, gap_scan_response event handler, gap_discover
response handler. Store discovered devices in BLED112Backend.discovered_devices.

+++ pygatt/pygatt.py
@@ -58,7 +58,8 @@ class BluetoothLEDevice(object):
                                            loglevel=LOG_LEVEL)
             self._mac_address = bytearray(
                 [int(b, 16) for b in mac_address.split(":")])
-            self._backend.delete_stored_bonds()
+            if delete_backend_bonds:
+                self._backend.delete_stored_bonds()
         elif backend == BACKEND['GATTTOOL']:
             raise NotImplementedError("TODO")
         else:
